created coffee route

+++ models/coffee.js
@@ -21,7 +21,7 @@ module.exports = (sequelize, DataTypes) => {
       name: { type: DataTypes.STRING, allowNull: false },
       shortDescription: { type: DataTypes.STRING, allowNull: false },
       longDescription: { type: DataTypes.TEXT, allowNull: false },
-      price: { type: DataTypes.INTEGER, allowNull: false },
+      price: { type: DataTypes.DECIMAL, allowNull: false },
       imageUrl: { type: DataTypes.STRING, allowNull: false },
     },
     {
